Refactor codebase to facilitate testing
