+++ colab_gemini_utils/prompts/example_prompt.md
