+++ pyproject.toml
@@ -1,0 +1,24 @@
+[project]
+name = "colab-gemini-utils"
+version = "0.1.0"
+description = "Helper functions for using Gemini as a coding assistant within Colab"
+authors = [
+    {name = "SamuelHLewis",email = "sam.lewis1@crowncommercial.gov.uk"}
+]
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "nbformat (>=5.10.4,<6.0.0)",
+    "pytest (>=8.4.2,<9.0.0)"
+]
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.poetry]
+name = "colab-gemini-utils"
+version = "0.1.0"
+packages = [
+    { include = "scripts", from = "colab_gemini_utils" }
+]
